Move schema to its own file
